+++ SemiBin/utils.py
@@ -1,12 +1,11 @@
 import os
 import subprocess
 from atomicwrites import atomic_write
-from Bio.Seq import Seq
-from Bio.SeqRecord import SeqRecord
-from Bio import SeqIO
 import sys
 import random
 import shutil
+
+from .fasta import fasta_iter
 
 def validate_args(args):
     def expect_file(f):
@@ -244,20 +243,15 @@ def write_bins(namelist, contig_labels, output, contig_dict,
 
     os.makedirs(output, exist_ok=True)
 
-    for label in res:
-        bin = []
-        whole_bin_bp = 0
-        for contig in res[label]:
-            rec = SeqRecord(
-                Seq(str(contig_dict[contig])), id=contig, description='')
-            bin.append(rec)
-            whole_bin_bp += len(str(contig_dict[contig]))
+    for label, contigs in res.items():
+        whole_bin_bp = sum(len(contig_dict[contig]) for contig in contigs)
 
-        ofname = f'bin.{label}.fa' if not recluster \
-                    else f'recluster_{origin_label}.bin.{label}.fa'
         if whole_bin_bp >= minfasta:
+            ofname = f'bin.{label}.fa' if not recluster \
+                    else f'recluster_{origin_label}.bin.{label}.fa'
             with atomic_write(os.path.join(output, ofname), overwrite=True) as ofile:
-                SeqIO.write(bin, ofile, 'fasta')
+                for contig in contigs:
+                    ofile.write(f'>{contig}\n{contig_dict[contig]}\n')
 
 
 def cal_kl(m, v):
@@ -352,33 +346,17 @@ def process_fasta(fasta_path, ratio):
     contig_length_list = []
     contig_dict = {}
 
-    for seq_record in SeqIO.parse(fasta_path, "fasta"):
-        if len(seq_record) >= 1000 and len(seq_record) <= 2500:
-            contig_bp_2500 += len(seq_record)
-        contig_length_list.append(len(seq_record))
-        whole_contig_bp += len(seq_record)
-        contig_dict[str(seq_record.id).strip('')] = str(seq_record.seq)
+    for h, seq in fasta_iter(fasta_path):
+        if 1000 <= len(seq) <= 2500:
+            contig_bp_2500 += len(seq)
+        contig_length_list.append(len(seq))
+        whole_contig_bp += len(seq)
+        contig_dict[h] = seq
 
     binned_short = contig_bp_2500 / whole_contig_bp < ratio
     must_link_threshold = get_must_link_threshold(contig_length_list)
     return binned_short, must_link_threshold, contig_dict
 
-def unzip_fasta(suffix, contig_path):
-    import gzip
-    import bz2
-    if suffix == 'gz':
-        contig_name = contig_path.replace(".gz", "")
-        ungz_file = gzip.GzipFile(contig_path)
-        open(contig_name, "wb+").write(ungz_file.read())
-        ungz_file.close()
-        return contig_name
-
-    if suffix == 'bz2':
-        contig_name = contig_path.replace(".bz2", "")
-        unbz2_file = bz2.BZ2File(contig_path)
-        open(contig_name, "wb+").write(unbz2_file.read())
-        unbz2_file.close()
-        return contig_name
 
 def split_data(data, sample, separator):
     """